Rename: Enum파일 수정
