+++ src/main/java/site/sonisori/sonisori/common/enums/Difficulty.java
@@ -1,4 +1,4 @@
-package site.sonisori.sonisori.common;
+package site.sonisori.sonisori.common.enums;
 
 public enum Difficulty {
 	EASY, MEDIUM, HARD
