+++ src/main/java/site/sonisori/sonisori/common/enums/Role.java
@@ -1,4 +1,4 @@
-package site.sonisori.sonisori.common;
+package site.sonisori.sonisori.common.enums;
 
 public enum Role {
 	ROLE_USER, ROLE_ADMIN
