+++ src/main/java/site/sonisori/sonisori/common/enums/SocialType.java
@@ -1,4 +1,4 @@
-package site.sonisori.sonisori.common;
+package site.sonisori.sonisori.common.enums;
 
 public enum SocialType {
 	google, kakao
